feat : viewmodel 사용 #7
