+++ app/src/main/java/org/android/go/sopt/presentation/login/LoginViewModel.kt
@@ -1,0 +1,52 @@
+package org.android.go.sopt.presentation.login
+
+import android.content.Intent
+import android.util.Log
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import org.android.go.sopt.R
+import org.android.go.sopt.SoptApplication
+import org.android.go.sopt.data.remote.ServicePool
+import org.android.go.sopt.data.remote.model.RequestSignInDto
+import org.android.go.sopt.data.remote.model.ResponseSignInDto
+import org.android.go.sopt.presentation.main.MainActivity
+import org.android.go.sopt.util.showShortToast
+import retrofit2.Call
+import retrofit2.Response
+
+class LoginViewModel: ViewModel() {
+    val id = MutableLiveData("")
+    val password = MutableLiveData("")
+
+    private val _signIn = MutableLiveData<ResponseSignInDto>()
+    val signIn: LiveData<ResponseSignInDto> get() = _signIn
+
+    fun signIn(){
+        ServicePool.signInService.signIn(
+            RequestSignInDto(
+                id.value.toString(), password.value.toString()
+            )
+        ).enqueue(object : retrofit2.Callback<ResponseSignInDto> {
+            override fun onResponse(
+                call: Call<ResponseSignInDto>,
+                response: Response<ResponseSignInDto>
+            ) {
+                if (response.isSuccessful) {
+                    SoptApplication.prefs.setBoolean(KEY_ISLOGIN, true)
+                    _signIn.value = response.body()
+                } else {
+                    Log.d("response err", response.body().toString())
+                }
+            }
+
+            override fun onFailure(call: Call<ResponseSignInDto>, t: Throwable) {
+                Log.d("errrr", t.message.toString())
+            }
+        })
+    }
+
+    companion object{
+        const val KEY_ISLOGIN = "isLogin"
+    }
+}
